Map Tiled qui marche !
! Attention la map est a exporter de façon manuelle !

+++ assets-cg/map/converters.py
@@ -26,10 +26,10 @@ def convert_map(input, output, params, target):
 	#structMap += fxconv.u32(h)
 	structMap += fxconv.u32(4*2**16) #startpos_x
 	structMap += fxconv.u32(4*2**16) #startpos_y
-    structMap += fxconv.u32(1*2**16) #startdir_x
+	structMap += fxconv.u32(1*2**16) #startdir_x
 	structMap += fxconv.u32(0)       #startdir_y
 	structMap += fxconv.u32(0)		 #startplane_x
-	structMap += fxconv.u32(0.66*2**16) #startplane_y
+	structMap += fxconv.u32(int(0.66*2**16)) #startplane_y
 
 	data_f = data["layers"][0]["data"] #Floor data
 	data_w = data["layers"][1]["data"] #Wall data
